Started implementation of deposits
